requirements.txt
Django==4.2.25
djangorestframework==3.16.1
drf-yasg==1.21.11
mysqlclient==2.2.7
python-decouple==3.8